requirements.txt
neo4j
pytest==8.3.4
flask==3.0.0
flask-cors==4.0.0
pyvis==0.3.2
networkx==3.2.1
python-dotenv==1.0.1
pandas==2.2.0



